Refactor pathing so roads are correctly seen as an improvement
With the default costs of 1 for plain, and when using 'auto' road cost
of `Math.ceil(1 / 2)`, this ends up being equal to the default plain
cost, leading to roads being utterly ignored.

Fix the issue by centralizing the defs and changing 'auto' mode to
give a slight bump to the other types' costs if they happen to have
the same value.