Protect against requests targeting inaccessible objects
This checks that targets for both LogisticsNetwork and 
TransportRequestGroup have at least one neighbor that's walkable,
fixing an issue that would cause CPU usage to shot up as pathing would
fail repeatedly.

Since this is only checking for neighborhood and not a reachable path,
this is not a complete fix.