Don't revert to normal queens once we can bunker
Once we have labs running, the normal queens are incapable of handling
requests coming from them, and if energy drops below the threshold the
colony starts falling apart as energy requests from the hatchery are not
processed anymore.

Remove the energy limit on whether bunker mode should be used or not to
stop reverting back to the normal queen overlord. Also name the two
overlords correctly so it's easier to tell which one is running, and
have them reassign queens that went through an overlord transition.