Have the SK reaper defense force be proportional to the invaders
I saw it spawn a lone hydralisk while a full fleet of 5 invaders was
wrecking the room. As a stop-gap, change the calculation to send as
many creeps as the invading force has. Ideally we would calculate the
actual amount of creeps needed to ensure victory, which I'd expect to be
less than that.