Overhaul miner overlord to get its behavior more consistent
I witnessed a drone spawn and instantly get to sleep as its affected
source was regenerating, blocking a spot in the middle of the bunker and
lowering efficiency by not being ready when the source was.

This merges back the many modes into one large function with appropriate
checks for sub-conditions (early/link/container/drop mining) which makes
it easier to have a shared behavior between all "modes".