Improve avoidDanger to try and stop the frozen creeps bug
This changes it so that it will ignore the colony room as a fallback if
it's the room under attack, forcing it to pick another colony nearby.
If it can't find any, then make the creep do a (more expensive) flee
from the hostiles, until we're out of the room.