Improve the settings loading
This makes all files use "config" to access settings, which merely
imports the contents of the settings file and reexports them cleanly.

It also adds a new DEFAULT_SETTINGS constant, which is what will be
used to generate the default Memory.settings object, allowing
player-specific control over it instead of hard-coding it.