Move the terminal/storage balancing into the command center
The goal is to be able to dispatch resources wherever they're needed
instead of the static terminal/storage task loop the manager did. This will
help fulfill in-flight requests from other parts of the bunker if there
are higher-priority requests for them at the same time (like from the
factory).

This also moves the "rebuilding terminal" quick-hauling of energy at
that level, since it's now only a tweak to the thresholds and
priorities.

The Resource Manager gains a new `targetForResource()` method, responsible
for checking any incoming resources for the thresholds and caps of the
storage and terminal, and returning the one that can still accept them.
The manager now makes use of this when moving resources around.