Fix an issue with managers getting stuck into a withdraw/deposit loop
What would happen is that if you'd get a lone low-priority
terminal-storage balancing request along with a higher priority supply
request from say the factory. The withdraw would be handled first, and
the high priority supply request for a non-matching resource would cause
the request matching to fail, with the manager having withdrawn. Then on
the next tick, with the withdraw request gone, the supply request would
be handled, causing the manager to transfer back its contents to make
room for the supply request's contents. And it would loop over until
something caused the priorities to shift slightly.

Fix the issue by special-casing the callback-based request search to not
stop at the first priority level, which causes the terminal-storage
matching to actually look at the rest of the requests, see its
counterpart and chain the tasks properly.