Fix a bug with the segmenter reparsing segments while they were modified
When updating the same segment data during the same tick, it would
end up losing any changes as the segment was parsed from its raw data
over and over. Change the cache check to use the cache if the mtime
is the same as the atime *and* we have already loaded that segment.