A couple improvements to transporter allocation
This adds a new per-remote priority for any extra transporters, so that
in the event of a cold-start, we don't keep spawning transporters for
sources that don't have a miner yet (since the owned room transporter
priority is lower than any of the remote room miners).

Additionally, tweak the transport power needed to only consider mining
sites that are active and have a miner, and in case of bootstrapping,
if they have a container or link that is somewhat full.