Fix expanding into owned rooms
As a side-effect of using `RoomIntel.isConsideredHostile`, which varies
with time, we could end up selecting a room that was already owned as
an expansion candidate, leading to unnecessary bloodbath.

Use the intel we have on the room to check that the room isn't owned by
a player already. Also add a global setting to tweak the minimum
distance from a non-allied player to consider as safe for expansion.