Make sure spawning directions are walkable
This adds a safety check to ensure that any spawn directions set from
requests are actually walkable. And also fixes a bug where as we checked
the core spawn's availability we would set the request's directions to
`[TOP, RIGHT]`, which would leak through to the following spawn in the
case it wasn't available, leading to stuck spawns.