Fix a bug where ranged creeps wouldn't be noticed
Since the BOOST object has final leaves that match neither a body part
nor a boost type, `getBodyPotential` would end up miscalculating the
potential of ranged creeps, leading to a complete failure of defenses.

This types boosts properly across the project, and closes the bug by
providing a map to convert a boost type to a modifier.